requirement.txt
autopep8==2.0.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pycodestyle==2.11.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
tomli==2.0.1
urllib3==2.0.4
